requirements.txt
numpy>=1.23
scipy>=1.10
geopandas>=0.13
shapely>=2.0
rasterio>=1.3
pyproj>=3.6
fiona>=1.9
torch>=2.0
scikit-image>=0.21
matplotlib>=3.7
pillow>=10.0
tqdm>=4.66
